requirements.txt
pip==23.2
click==8.0.4
colorama==0.4.6
flask==2.2.2
gunicorn==21.2.0
itsdangerous==2.0.1
jinja2==3.1.2
joblib==1.2.0
markupsafe==2.1.1
numpy==1.25.0
numpy-base==1.25.0
pandas==1.5.3
pillow==9.4.0
pytorch==2.0.1
pytorch-cuda==11.8
pytorch-lightning==2.0.4
pytorch-model-summary==0.1.1
pytorch-mutex==1.0
torchaudio==2.0.2
torchmetrics==0.11.4
torchsummary==1.5.1
torchvision==0.15.2
tornado==6.2
werkzeug==2.2.3
wtforms==2.3.3